+++ src/background/seelen_bar/hook.rs
@@ -1,24 +1,33 @@
-use windows::Win32::{
-    Foundation::HWND,
-    UI::WindowsAndMessaging::{
-        EVENT_OBJECT_FOCUS, EVENT_OBJECT_NAMECHANGE, EVENT_SYSTEM_FOREGROUND,
-    },
-};
+use windows::Win32::Foundation::HWND;
 
-use crate::error_handler::Result;
+use crate::{error_handler::Result, windows_api::WindowsApi, winevent::WinEvent};
 
 use super::FancyToolbar;
 
 impl FancyToolbar {
-    pub fn process_win_event(&mut self, event: u32, hwnd: HWND) -> Result<()> {
+    pub fn process_win_event(&mut self, event: WinEvent, hwnd: HWND) -> Result<()> {
         match event {
-            EVENT_OBJECT_NAMECHANGE => {
+            WinEvent::ObjectNameChange => {
                 if self.last_focus == Some(hwnd.0) {
                     self.focus_changed(hwnd)?;
                 }
             }
-            EVENT_SYSTEM_FOREGROUND | EVENT_OBJECT_FOCUS => {
+            WinEvent::SystemForeground | WinEvent::ObjectFocus => {
                 self.focus_changed(hwnd)?;
+            }
+            WinEvent::SyntheticFullscreenStart => {
+                let monitor = WindowsApi::monitor_from_window(self.window.hwnd()?);
+                let event_monitor = WindowsApi::monitor_from_window(hwnd);
+                if monitor == event_monitor {
+                    self.hide()?;
+                }
+            }
+            WinEvent::SyntheticFullscreenEnd => {
+                let monitor = WindowsApi::monitor_from_window(self.window.hwnd()?);
+                let event_monitor = WindowsApi::monitor_from_window(hwnd);
+                if monitor == event_monitor {
+                    self.show()?;
+                }
             }
             _ => {}
         };